refactor: move spacecraft.py into new spacecraft folder
